common lisp implementation
Working common lisp implementation of the algorithm. This will compute
the scores in a multi-threaded manner (not sure how to get into
distributed computing yet--bordeaux-threads may not have an answer for
that kind of work). It could be improved, but my precursory benchmarks
have outperformed julia! Not a perfect test because I'm using manjaro
instead of mint (newer kernel), but in broad strokes the results seem
very promising.