using unix time instead of formatted times, only awk used is gawk
I wanted more accurate results in the analysis code, so the progress
logs now print the unix time with microsecond resolution rather than a
formatted datetime. This led to a change in dependencies for some
languages.
different awk implementations are kind of annoying to install in a
container, and I wanted to use the gettimeofday function from gawk's
time library, so gawk is the only version of awk being used now.