implementation of database to compute all outcomes (impractical)
r/construct_database.R changes the algorithm to compute and store
every outcome along the tree of guesses. rather than computing the
score for the opening word, the values that comprise the 'steps'
to compute these scores are stored in a database, and the algorithm
to compute the states of subsequent guesses is also implemented
(the game of wordle can be thought of as a tree where each opening
guess and color combo is a root node and the following guesses fan
out from under it). this implementation is currently impractical
with my hardware since the second guess alone would have around 40
billion outcomes with even more in further guesses in the game. Some
preliminary guesses indicate that the second guesses would take 450
days to compute on my hardware (ignoring other hardware constraints
such as disk space, which would also run out much sooner than this).
It is a reasonable guess to say that there are over a trillion nodes
in this tree. So, without serious optimization and dishing out
hundreds of dollars to a cloud provider, this 'compute every node
in the tree of guesses' approach is dead in the water. I already
have the solution but have not yet staged it in git since I think
that picking up on this recursive algorithm is an interesting
opportunity to tinker with large compute clusters. Therefore, I'm
staging the impractical solution here and can revisit this commit
somewhere down the road. I may never do it, but I don't want to burn
any bridges with my future myself.

app.R requires a browser to be set, so I set one.
I don't remember the purpose of r/play.R's changes, but
they're going to be replaced soon enough anyhow.
r/test.R will likewise be replaced soon enough.