using postgres for its NUMERIC type
building the database in postgres and deploying it using aiven.
this schema still needs adjustment because the values do not store
the total frequency of the words that comprise the set. these values could be
computed, but the aiven deployment restricts the database to a single core with
1GB RAM, so resource limitations require the totals to be precomputed.
perhaps this wouldn't be as big of an issue for an embedded database, although I
could see a similar but less severe issue happening on the server that
deploys the shiny app.

regardless, the plan for future commits is to include the frequency
total of the words that comprise the remaining_words_set_id and store it
as a column in guess_combo_info. then, the shiny app will be able to
compute scores without needing to calculate the frequency total of every
word in the set, as it can just look up a given word and then subtract
the values that do not intersect with it??? still thinking this one
through--the goal is to rapidly compute the top 10 highest-scoring words
among the remaining words in the set as a suggestion to the user in
the app.